Refactoring + added exception handling

+++ design/src/main/java/model/discount/Discount.java
@@ -3,6 +3,8 @@ package model.discount;
 
 import model.discount.discountrule.DiscountRule;
 
+import java.math.BigDecimal;
+
 /**
  * A model representing a discount that is applied to
  * a specific sale
@@ -11,7 +13,7 @@ public interface Discount{
     void setTotalPriceReduction(double totalPriceReduction);
     Number getRequirement();
     DiscountRule getDiscountRule();
-    Double getTotalPriceReduction();
+    BigDecimal getTotalPriceReduction();
     String getDate();
     boolean isDefault();
 }
